requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
chromadb==0.5.0
sentence-transformers==2.7.0
openai==1.30.1
pdfplumber==0.11.0
python-docx==1.1.2
pandas==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1
